requirements.txt
asttokens==3.0.0
attrs==25.1.0
Automat==24.8.1
boto3==1.37.6
botocore==1.37.6
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
colorama==0.4.6
comm==0.2.2
constantly==23.10.4
cryptography==44.0.2
cssselect==1.2.0
debugpy==1.8.13
decorator==5.2.1
defusedxml==0.7.1
executing==2.2.0
fastjsonschema==2.21.1
filelock==3.17.0
googlemaps==4.10.0
greenlet==3.1.1
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
ipykernel==6.29.5
ipython==9.0.2
ipython_pygments_lexers==1.1.1
itemadapter==0.11.0
itemloaders==1.3.2
jedi==0.19.2
jmespath==1.0.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
lxml==5.3.1
matplotlib-inline==0.1.7
mysql-connector-python==9.2.0
narwhals==1.30.0
nbformat==5.10.4
nest-asyncio==1.6.0
numpy==2.2.3
packaging==24.2
pandas==2.2.3
parsel==1.10.0
parso==0.8.4
platformdirs==4.3.6
playwright==1.50.0
plotly==6.0.0
prompt_toolkit==3.0.50
Protego==0.4.0
psutil==7.0.0
psycopg2==2.9.10
psycopg2-binary==2.9.10
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyDispatcher==2.0.7
pyee==12.1.1
Pygments==2.19.1
pyOpenSSL==25.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.1
pywin32==309
pyzmq==26.2.1
queuelib==1.7.0
referencing==0.36.2
requests==2.32.3
requests-file==2.1.0
rpds-py==0.23.1
s3transfer==0.11.4
Scrapy==2.12.0
scrapy-playwright==0.0.43
service-identity==24.2.0
setuptools==75.8.2
six==1.17.0
SQLAlchemy==2.0.38
stack-data==0.6.3
tldextract==5.1.3
tornado==6.4.2
traitlets==5.14.3
Twisted==24.11.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
w3lib==2.3.1
wcwidth==0.2.13
zope.interface==7.2
